feat: :sparkles: Users - i18n - configuration - cloudinary
multilanguage implementation with the i18n module in the feature and shared modules, in addition to their respective DTOs. Changes in the code configuration, now dividing it by namespaces. Implementation of the "cloudinary" shared type module for managing files on a CDN. Creation of the featured "users" type module, with its respective endpoints to manage users.